# bitset operations are aware of deletes