--- a/package.json
+++ b/package.json
@@ -39,10 +39,10 @@
     "typedfastbitset": "~0.2.1"
   },
   "peerDependencies": {
-    "async-append-only-log": "^4.2.1"
+    "async-append-only-log": "ssbc/async-append-only-log#deleted-obz"
   },
   "devDependencies": {
-    "async-append-only-log": "^4.3.1",
+    "async-append-only-log": "ssbc/async-append-only-log#deleted-obz",
     "expose-gc": "^1.0.0",
     "flumecodec": "0.0.1",
     "flumelog-offset": "3.4.4",