--- a/test/del.js
+++ b/test/del.js
@@ -2,13 +2,13 @@
 //
 // SPDX-License-Identifier: Unlicense
 
-const test = require('tape')
 const validate = require('ssb-validate')
 const ssbKeys = require('ssb-keys')
 const path = require('path')
-const { prepareAndRunTest, addMsg, helpers } = require('./common')()
+const pify = require('util').promisify
 const rimraf = require('rimraf')
 const mkdirp = require('mkdirp')
+const { prepareAndRunTest, addMsgPromise, helpers } = require('./common')()
 
 const dir = '/tmp/jitdb-add'
 rimraf.sync(dir)
@@ -16,16 +16,16 @@ mkdirp.sync(dir)
 
 var keys = ssbKeys.loadOrCreateSync(path.join(dir, 'secret'))
 
-prepareAndRunTest('Delete', dir, (t, db, raf) => {
-  const msg1 = { type: 'post', text: 'Testing 1' }
-  const msg2 = { type: 'post', text: 'Testing 2' }
-  const msg3 = { type: 'post', text: 'Testing 3' }
+prepareAndRunTest('delete then index', dir, async (t, jitdb, log) => {
+  const content1 = { type: 'post', text: 'Testing 1' }
+  const content2 = { type: 'post', text: 'Testing 2' }
+  const content3 = { type: 'post', text: 'Testing 3' }
   let state = validate.initial()
-  state = validate.appendNew(state, null, keys, msg1, Date.now())
-  state = validate.appendNew(state, null, keys, msg2, Date.now() + 1)
-  state = validate.appendNew(state, null, keys, msg3, Date.now() + 2)
+  state = validate.appendNew(state, null, keys, content1, Date.now())
+  state = validate.appendNew(state, null, keys, content2, Date.now() + 1)
+  state = validate.appendNew(state, null, keys, content3, Date.now() + 2)
 
-  const typeQuery = {
+  const query = {
     type: 'EQUAL',
     data: {
       seek: helpers.seekType,
@@ -35,28 +35,65 @@ prepareAndRunTest('Delete', dir, (t, db, raf) => {
     },
   }
 
-  addMsg(state.queue[0].value, raf, (err, msg1) => {
-    addMsg(state.queue[1].value, raf, (err, msg2, offset2) => {
-      addMsg(state.queue[2].value, raf, (err, msg3) => {
-        raf.del(offset2, () => {
-          db.paginate(
-            typeQuery,
-            0,
-            10,
-            false,
-            false,
-            'declared',
-            (err, { results }) => {
-              t.deepEqual(results, [msg1, msg3])
+  const msg1 = (await addMsgPromise(state.queue[0].value, log)).msg
+  const offset2 = (await addMsgPromise(state.queue[1].value, log)).offset
+  const msg3 = (await addMsgPromise(state.queue[2].value, log)).msg
 
-              db.all(typeQuery, 0, false, false, 'declared', (err, results) => {
-                t.deepEqual(results, [msg1, msg3])
-                t.end()
-              })
-            }
-          )
-        })
-      })
-    })
-  })
+  await pify(log.del)(offset2)
+  await pify(log.onDeletesFlushed)()
+
+  const answer = await pify(jitdb.paginate)(
+    query,
+    0,
+    2,
+    false,
+    false,
+    'declared'
+  )
+  t.deepEqual(answer.results, [msg1, msg3])
+
+  const results = await pify(jitdb.all)(query, 0, false, false, 'declared')
+  t.deepEqual(results, [msg1, msg3])
+})
+
+prepareAndRunTest('index then delete', dir, async (t, jitdb, log) => {
+  const content1 = { type: 'post', text: 'Testing 1' }
+  const content2 = { type: 'post', text: 'Testing 2' }
+  const content3 = { type: 'post', text: 'Testing 3' }
+  let state = validate.initial()
+  state = validate.appendNew(state, null, keys, content1, Date.now())
+  state = validate.appendNew(state, null, keys, content2, Date.now() + 1)
+  state = validate.appendNew(state, null, keys, content3, Date.now() + 2)
+
+  const query = {
+    type: 'EQUAL',
+    data: {
+      seek: helpers.seekType,
+      value: helpers.toBipf('post'),
+      indexType: 'type',
+      indexName: 'type_post',
+    },
+  }
+
+  const msg1 = (await addMsgPromise(state.queue[0].value, log)).msg
+  const offset2 = (await addMsgPromise(state.queue[1].value, log)).offset
+  const msg3 = (await addMsgPromise(state.queue[2].value, log)).msg
+
+  await pify(jitdb.prepare)(query)
+
+  await pify(log.del)(offset2)
+  await pify(log.onDeletesFlushed)()
+
+  const answer = await pify(jitdb.paginate)(
+    query,
+    0,
+    2,
+    false,
+    false,
+    'declared'
+  )
+  t.deepEqual(answer.results, [msg1, msg3])
+
+  const results = await pify(jitdb.all)(query, 0, false, false, 'declared')
+  t.deepEqual(results, [msg1, msg3])
 })