Draft PR for issue #223

@arj03 Check the approach I'm using here, where I listen to an obz from AAOL. See also the AAOL branch https://github.com/ssbc/async-append-only-log/pull/82

Before proceeding with polishing this idea, I want to see what you think and I want to check the performance.