@arj03 2022-07-14
I guess we can try this approach, we still need to handle that there might be deletes we did not get because this way of signalling deletes can fail (see my comment here: https://github.com/ssbc/jitdb/issues/223). I'm also interested in the seeing the performance aspect of this bitset approach, that is at least something (as you write) to test now :)