@staltz 2022-07-14
> this way of signalling deletes can fail

How? (I read the comment but I didn't understand how it could fail)