@staltz 2022-07-14
Notice the benchmark:

[**Before:**](https://github.com/ssbc/jitdb/pull/222)

| | |
|---|---|
| Paginate 20000 msgs with pageSize=5 | 8093.26ms ± 509.73ms |

**After:**

| | |
|---|---|
| Paginate 20000 msgs with pageSize=5 | 17007.6ms ± 250.82ms |