@arj03 2022-07-15
> Notice the benchmark:
> 
> [**Before:**](https://github.com/ssbc/jitdb/pull/222)
> 	
> Paginate 20000 msgs with pageSize=5 	8093.26ms ± 509.73ms
> 
> **After:**
> 	
> Paginate 20000 msgs with pageSize=5 	17007.6ms ± 250.82ms
> 
> All the other numbers seem to be okayish.

I put in a timing on the `intersectWithPresent` and its around 0.03 - 0.3 ms. The problem with this paginate is that you run this soo many times (4000).