@staltz 2022-07-16
I'm bothered about the deeper issues in this issue, and I thought "do we have to solve this?".

There are two problems going on:

- duplication of items when paginating
- pages shorter than what was requested as the page size

Problem 1 is a big deal. Problem 2 is a small deal. So maybe we don't need to solve number 2, we just need to make sure that number 1 is solved. And I think that might not be too hard, just gotta track the seqs better. I'll try building that soon.